Fixes a problem with Windows file separator in Fixt
File.separator is used in a regex, but the Windows filesystem separator
is a backslash which is an escape character in a regex. There is no
subsequent character and that causes an exception.

Redid Fixt.java as Fixt.groovy as it's in the src/main/groovy folder.

Switched to using FileSystems.default.getPath to resolve the location.
Sidesteps the issue completely, revert to the old code to see an example
of how it fails.

Excluded UnnecessaryGString from Codenarc configuration as there is no
such thing as unnecessary gstring. Double quoted strings are still just
java.lang.String if there's no interpolation in them.  Ref: 4.4 of the
groovy docs.

Extended the spec to test with both a unix and a windows file separator